requirements.txt
blessed==1.20.0
bpython==0.24
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
curtsies==0.4.2
cwcwidth==0.1.9
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
networkx==3.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyxdg==0.28
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
trans==2.1.0
urllib3==2.2.2
wcwidth==0.2.13
